+++ src/test/elements/allbound/leads.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const common = require('../common');
+
+describe('leads', () => {
+  common.for('marketing', 'leads');
+});