+++ src/test/elements/amazons3/files.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const common = require('../common');
+
+describe('files', () => {
+  common.for('documents', 'files');
+});