+++ src/test/elements/autotaskhelpdesk/tasks.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const common = require('../common');
+
+describe('users', () => {
+  common.for('helpdesk', 'users');
+});