+++ src/test/elements/box/folders.js
@@ -1,15 +1,17 @@
 'use strict';
 
-const util = require('util');
 const chakram = require('chakram');
 const expect = chakram.expect;
-const url = require('url');
+const common = require('../common');
 
 describe('folders', () => {
+  common.for('crm','folders');
+
   it('should allow listing folder contents', () => {
     var uri = '/hubs/documents/folders/contents';
-    return chakram.get(uri + '?path=/').then((r) => {
-      expect(r).to.have.status(200);
-    });
+    return chakram.get(uri + '?path=/')
+      .then(r => {
+        expect(r).to.have.status(200);
+      });
   });
 });