+++ src/test/elements/closeio/contacts.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const common = require('../common');
+
+describe('contacts', () => {
+  common.for('crm', 'contacts');
+});