+++ src/test/elements/docusign/envelopes.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const common = require('../common');
+
+describe('envelopes', () => {
+  common.for('esignature', 'envelopes', 'foobarid');
+});