+++ src/test/elements/etsy/customers.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const common = require('../common');
+
+describe('customers', () => {
+  common.for('ecommerce', 'customers');
+});