+++ src/test/setup.js
@@ -3,7 +3,7 @@
 const chakram = require('chakram');
 const util = require('util');
 const argv = require('optimist').demand('user').argv
-const props = require('core/util/properties');
+const props = require('core/props');
 
 before((done) => {
   const user = argv.user;
@@ -19,10 +19,10 @@ before((done) => {
   const form = { j_username: user, j_password: password };
 
   chakram.post(secUrl, null, { jar: true, form: form })
-    .then((r) => {
+    .then(r => {
       return chakram.get(url + '/elements/api-v1/ui/getSecrets', { jar: true });
     })
-    .then((r) => {
+    .then(r => {
       props.set('user.secret', r.body.user);
       props.set('org.secret', r.body.company);
       chakram.setRequestDefaults({
@@ -33,7 +33,7 @@ before((done) => {
       });
       done();
     })
-    .catch((r) => {
+    .catch(r => {
       console.log('Well shucks...failed to finish setup...\n  Is %s up and running?\n  Do you have the write username and password?', url);
       process.exit(1);
     });